Documentation fix for export endpoint permissions
If you want to export the FPF format findings via the `/api/v1/finding/project/{uuid}/export` API, the documentation says that you need the VULNERABILITY_ANALYSIS permission for the accompanying API key. After some debugging, i found out that the source requires VIEW_VULNERABILITY:

https://github.com/DependencyTrack/dependency-track/blob/6666828756543de1056461fb20f8dbd5ac3ebe57/src/main/java/org/dependencytrack/resources/v1/FindingResource.java#L111

This is a fix to patch the documentation to state the correct permissions needed.

Signed-off-by: Tonimir Kisasondi <kisasondi@gmail.com>