Add affected project in email subject
Shows the affected project in the subject line of an email notification
to make the emails more distinguishable and to make it easier for a user
to gather important information like project name and version quickly.

Changes the default email template for the notification groups
'NEW_VULNERABILITY_IDENTIFIED' and 'PROJECT_AUDIT_CHANGE' to list the
project of the affected component and to list other projects, if these
are also affected.

Signed-off-by: RBickert <rbt@mm-software.com>