Ensure that policy is part of policy violations API response
DataNucleus would (in some, not all cases) falsely assume that the `violation.policyCondition.policy` field has already been loaded, and would not attempt to fetch it again. This then caused API responses to not contain the `policy` field (#2043).

The scenario is not reproducible in unit tests, even with regular cache eviction. It is reliably reproducible in running instances using the steps listed in #2043 though.

Instead of detachment like it's introduced in this PR, another cheap fix would've been to call `violation.getPolicyCondition().getPolicy().getName()` instead of `violation.getPolicyCondition().getPolicy()` in `PolicyQueryManager#getPolicyViolations`. That would force DN to fetch all fields of `Policy`, but I felt like that was too hacky to rely on it.

Signed-off-by: nscuro <nscuro@protonmail.com>