Disable DataNucleus L2 cache for mirroring tasks
Objects that are persisted (created, updated) are put into DN's second level cache. Per default, the cache keeps soft references to objects.

https://www.datanucleus.org/products/accessplatform_6_0/jdo/persistence.html#cache_level2

While this kind of cache may be beneficial for BAU operations, it is totally overkill for our mirroring tasks. Over the course of mirroring the NVD for the first time alone, the cache will grow to over 2 million objects. The majority of those objects will never be accessed again, especially not soon enough for the cache to be of any use.

Soft references can potentially stay for a long time, and thus may inflate heap utilization.

There is a bug in DN that caused NPEs in `Vulnerability#setCwes` on newly created `Vulnerability` objects when L2 cache is disabled. The DN-internal method `dnGetcwes` would return `null`, despite the `cwes` field being initialized with `new ArrayList<>()` one line before. Not sure why this happens, but it can be prevented by setting the `cwes` field to a fully initialized list.

Signed-off-by: nscuro <nscuro@protonmail.com>