Add Docker Compose files for simplified local testing
These make it easier to test database migrations and compatibility with the different RDBMSes in general. Also simplifies testing the impact of performance critical functions with Prometheus metrics.

But to be honest the main motivation to add these files is that I always forget how to set up MSSQL locally.

Signed-off-by: nscuro <nscuro@protonmail.com>