Use `exec` to launch JVM process in container
`java` is currently executed from within a shell (default `ENTRYPOINT` of the base image). This is required because the `CMD` makes use of variable substitution not available outside a shell.

The problem is that in doing so, the JVM process does not receive a `SIGTERM` signal when the container runtime stops the container (e.g. via `docker stop`). This prevents Dependency-Track from performing a graceful shutdown.

See also:
* https://stackoverflow.com/a/31840306
* https://aws.amazon.com/blogs/containers/graceful-shutdowns-with-ecs/

Credit to @valentijnscholten for discovering this. https://owasp.slack.com/archives/C6R3R32H4/p1684243000301239

Signed-off-by: nscuro <nscuro@protonmail.com>