Prevent `IndexEvent` from being singleton
As `IndexEvent`s are being handled in the single-threaded event service, they'll be handled serially. Dispatching them multiple times is mostly unproblematic.

The logic in `SearchManager#reindex` is not working as expected when the event is a singleton, because chain identifiers are inherited from the first event in the chain, causing #2987.

Fixes #2987

Signed-off-by: nscuro <nscuro@protonmail.com>