Fix impossible SQL query conditions causing DB indexes to be bypassed
JDOQL conditions like `field != null && field == :field` can result in the SQL `FIELD IS NOT NULL AND FIELD IS NULL` being generated when the `:field` parameter is `null`. Such SQL condition cause the database's query planner to be unable to use any indexes on affected columns.

This is now prevented by only including fields in the query if the parameter is not `null`.

Additionally, fixed another occurrence of #2540.

Thanks to Mikael Carneholm for reporting this on Slack: https://owasp.slack.com/archives/C6R3R32H4/p1698071932430299

Signed-off-by: nscuro <nscuro@protonmail.com>