Fix NVD API's last modified timestamp requiring restart to be applied
The timestamp is created with value `null` in a context where L2 caching is enabled (`DefaultObjectGenerator`), but updated in one where L2 caching is disabled (`NistApiMirrorTask#updateLastModified`). What's more, it's queried in a context with L2 caching enabled (`NistApiMirrorTask#inform`). This causes the property to be updated, but the cache not being invalidated, thus still serving queries.

To not have similar inconsistencies in other areas of the application, re-enable L2 caching when updating the property.

Eventually, L2 caching should be disabled system-wide.

Signed-off-by: nscuro <nscuro@protonmail.com>