Fix unique constraint violation during NVD mirroring
The fix is achieved by using the same logic for persisting `Vulnerability` and `VulnerableSoftware` records that `NistApiMirrorTask` was already using. It handles duplicate records.

This should also yield a performance boost (did not benchmark because that wasn't the focus of this change), since the transaction commit frequency is reduced compared to the previous logic.

Fixes #3663

Signed-off-by: nscuro <nscuro@protonmail.com>