Reduce preemptive expensive calls in `ResourceTest`
Turns out that hashing the password for the `testUser` that is created in `ResourceTest` takes >=1sec, making it the biggest contributing factor to test duration. Instead of re-hashing the same string over and over again, hash it once and re-use where appropriate.

The loading of default objects was another factory. Some tests only need parts of the default objects, not all. Some are more expensive to load than others (loading >700 licenses takes longer than ~10 permissions).

Tests that previously took over a second to run now complete in ~50ms.

Signed-off-by: nscuro <nscuro@protonmail.com>