Disable DataNucleus L2 cache globally
Currently, DataNucleus will put all objects into the L2 cache. Given the volume of objects being processed by DT, this behavior quickly adds up to enormous cache sizes. Users continue to be bamboozled by DT's memory requirements, which for a large part are driven by the wasteful L2 caching.

While working on https://github.com/DependencyTrack/dependency-track/pull/4305, it became obvious that the hit rates of the cache are absolutely dwarfed by the high rate of misses. Storing such large volumes of objects in RAM is simply not justified if hit rates are that low.

Disabling the L2 cache solves a lot of recurring issues we and users are facing. If we want to introduce caching again in the future, we should do it in targeted areas, and preferably not directly in the persistence layer.

We disabled the L2 cache in Hyades a long time ago, and it has worked out very well for us. It was a precondition to make the API server horizontally scalable. Some more context:

* https://github.com/DependencyTrack/hyades/issues/375#issuecomment-1535929250
* https://github.com/DependencyTrack/hyades/issues/576

Supersedes #4305

Signed-off-by: nscuro <nscuro@protonmail.com>