Adjust `null` handling of `collectionLogic` and `collectionLogicChanged`
Store `ProjectCollectionLogic.NONE` as `NULL`. Reason being that the vast majority of projects will not have a collection logic, and storing `NULL` is cheaper than storing the string `NONE` over and over again. The code and REST API will still treat `NULL` as `NONE`, though.

Remove the index on `Project.collectionLogic`. The column has a low cardinality and is only ever queried in `PortfolioMetricsUpdateTask#fetchNextActiveProjectsBatch`. Since this method paginates using the `id` column, the index on `collectionLogic` is never used.

Remove the redundant upgrade item since `collectionLogic` columns are expected to remain `NULL` when a project is not a collection.

Make `ProjectMetrics.collectionLogicChanged` non-nullable and default to `false`. Prevents tri-state logic.

Signed-off-by: nscuro <nscuro@protonmail.com>